+++ tests/example_for_tests/Untiltelt.py
@@ -1,0 +1,1 @@
+from potential import test_potential